monitor-web-starter/src/integrationTest/java/org/bambrikii/monitoring/envminidashboard/web/DashboardControllerTest.java
package org.bambrikii.monitoring.envminidashboard.web;

import org.bambrikii.monitoring.envminidashboard.model.Dashboard;
import org.bambrikii.monitoring.envminidashboard.web.config.EnvMonDashboardConfig;
import org.bambrikii.monitoring.envminidashboard.web.controllers.DashboardController;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.orm.jpa.AutoConfigureDataJpa;
import org.springframework.boot.test.autoconfigure.web.client.AutoConfigureWebClient;
import org.springframework.boot.test.autoconfigure.web.client.RestClientTest;
import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
import org.springframework.http.ResponseEntity;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.context.web.WebAppConfiguration;
import org.springframework.test.web.client.MockRestServiceServer;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;
import static org.springframework.test.web.client.match.MockRestRequestMatchers.requestTo;
import static org.springframework.test.web.client.response.MockRestResponseCreators.withSuccess;

@Transactional
@ContextConfiguration(classes = {
        EnvMonDashboardConfig.class
})
@AutoConfigureDataJpa
@RunWith(SpringRunner.class)
@WebAppConfiguration
@AutoConfigureMockMvc
@RestClientTest(DashboardController.class)
@AutoConfigureWebClient(registerRestTemplate = true)
public class DashboardControllerTest {
    @Autowired
    private DashboardController client;

    @Autowired
    private MockRestServiceServer server;

    @Test
    public void shouldRespond() throws Exception {
        this.server
                .expect(requestTo("/env-monitor/dashboard"))
                .andRespond(withSuccess())
        ;
    }

    @Test
    public void shouldDashboardAddEnvs() {
        assertThat(client).isNotNull();

        ResponseEntity<List<Dashboard>> dashboard = client.list();
        assertThat(dashboard).isNotNull();
    }

}
